+++ ts/examples/tools/src/tool-conversion.ts
@@ -1,11 +1,12 @@
 import { Composio, jsonSchemaToZodSchema } from '@composio/core';
+import { nullable, z } from 'zod';
 import zodToJsonSchema from 'zod-to-json-schema';
 
 const composio = new Composio({
   apiKey: process.env.COMPOSIO_API_KEY,
 });
 
-const tool = await composio.tools.getRawComposioToolBySlug('COMPOSIO_MULTI_EXECUTE_TOOL');
+const tool = await composio.tools.getRawComposioToolBySlug('SLACK_SEND_MESSAGE');
 const inputParams = tool.inputParameters;
 console.log(`---------- Original Input Parameters ----------`);
 console.log(JSON.stringify(inputParams, null, 2));