+++ ts/packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @composio/core
 
+## 0.1.51
+
+### Patch Changes
+
+- b9b005a: Add support for composio connect links
+
 ## 0.1.50
 
 ### Patch Changes